requirements.txt
python-telegram-bot==13.7
requests==2.28.1
pandas==2.0.0
ta==0.10.2
web3==6.0.0
python-dotenv==0.21.0
